requirements.txt
fastapi
uvicorn
google-generativeai
pydantic
python-multipart
requests
python-dotenv
